matrix inversion converted to HLS 2021.1 and adding iir filter

+++ mat_inv/src/inv_4x4.hpp
@@ -2,7 +2,7 @@
 //#include <vector>
 using namespace std;
 
-typedef double CF_base;
+typedef float CF_base;
 typedef std::complex<CF_base> CF ;
 
 void inv_4x4(CF x[4][4], CF y[4][4]);
